student grade exercise working

+++ src/shapes/Measurable.java
@@ -2,7 +2,7 @@ package shapes;
 
 public interface Measurable {
 
-    double getPerimeter();
-    double getArea();
+    int getPerimeter();
+    int getArea();
 
 }
